requirements.txt
cycler==0.10.0
decorator==4.4.2
deskew==0.10.32
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
opencv-python==4.5.3.56
pandas==1.3.0
pdf2image==1.16.0
Pillow==8.3.1
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.2
typing==3.7.4.3
